ofdm_get_test_bits working in and uncoded modes
